pyproject.toml
[project]
name = "ckart-cli"
version = "0.1"
description = "ckart CLI to manage VMs"
authors = [
    { name = "Project ckart", email = "hetarth.jain@gmail.com" }
]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.7"
dependencies = [
    "requests",
    "python-dotenv",
    "tabulate",
    "pynacl",
    "psutil"
]

[project.scripts]
ckart = "ckart.main:main"

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"
